backend/requirements-dev.txt
pytest
pytest-asyncio
httpx
respx
pytest-mock
fakeredis
aioredis
sqlalchemy
alembic
uvicorn
coverage
black
ruff